Struct comrak::arena_tree::Node

pub struct Node<'a, T: 'a> {
    pub data: T,
    // some fields omitted
}

A node inside a DOM-like tree.

Fields
data: T

The data held by the node.

Methods

impl<'a, T> Node<'a, T>

pub fn new(data: T) -> Node<'a, T>

Create a new node from its associated data.

Typically, this node needs to be moved into an arena allocator before it can be used in a tree.

pub fn parent(&self) -> Option<&'a Node<'a, T>>

Return a reference to the parent node, unless this node is the root of the tree.

pub fn first_child(&self) -> Option<&'a Node<'a, T>>

Return a reference to the first child of this node, unless it has no child.

pub fn last_child(&self) -> Option<&'a Node<'a, T>>

Return a reference to the last child of this node, unless it has no child.

pub fn previous_sibling(&self) -> Option<&'a Node<'a, T>>

Return a reference to the previous sibling of this node, unless it is a first child.

pub fn next_sibling(&self) -> Option<&'a Node<'a, T>>

Return a reference to the previous sibling of this node, unless it is a last child.

pub fn same_node(&self, other: &Node<'a, T>) -> bool

Returns whether two references point to the same node.

Important traits for Ancestors<'a, T>
impl<'a, T> Iterator for Ancestors<'a, T> type Item = &'a Node<'a, T>;

pub fn ancestors(&'a self) -> Ancestors<'a, T>

Return an iterator of references to this node and its ancestors.

Call .next().unwrap() once on the iterator to skip the node itself.

Important traits for PrecedingSiblings<'a, T>
impl<'a, T> Iterator for PrecedingSiblings<'a, T> type Item = &'a Node<'a, T>;

pub fn preceding_siblings(&'a self) -> PrecedingSiblings<'a, T>

Return an iterator of references to this node and the siblings before it.

Call .next().unwrap() once on the iterator to skip the node itself.

Important traits for FollowingSiblings<'a, T>
impl<'a, T> Iterator for FollowingSiblings<'a, T> type Item = &'a Node<'a, T>;

pub fn following_siblings(&'a self) -> FollowingSiblings<'a, T>

Return an iterator of references to this node and the siblings after it.

Call .next().unwrap() once on the iterator to skip the node itself.

Important traits for Children<'a, T>
impl<'a, T> Iterator for Children<'a, T> type Item = &'a Node<'a, T>;

pub fn children(&'a self) -> Children<'a, T>

Return an iterator of references to this node’s children.

Important traits for ReverseChildren<'a, T>
impl<'a, T> Iterator for ReverseChildren<'a, T> type Item = &'a Node<'a, T>;

pub fn reverse_children(&'a self) -> ReverseChildren<'a, T>

Return an iterator of references to this node’s children, in reverse order.

Important traits for Descendants<'a, T>
impl<'a, T> Iterator for Descendants<'a, T> type Item = &'a Node<'a, T>;

pub fn descendants(&'a self) -> Descendants<'a, T>

Return an iterator of references to this node and its descendants, in tree order.

Parent nodes appear before the descendants. Call .next().unwrap() once on the iterator to skip the node itself.

Important traits for Traverse<'a, T>
impl<'a, T> Iterator for Traverse<'a, T> type Item = NodeEdge<&'a Node<'a, T>>;

pub fn traverse(&'a self) -> Traverse<'a, T>

Return an iterator of references to this node and its descendants, in tree order.

Important traits for ReverseTraverse<'a, T>
impl<'a, T> Iterator for ReverseTraverse<'a, T> type Item = NodeEdge<&'a Node<'a, T>>;

pub fn reverse_traverse(&'a self) -> ReverseTraverse<'a, T>

Return an iterator of references to this node and its descendants, in tree order.

pub fn detach(&self)

Detach a node from its parent and siblings. Children are not affected.

pub fn append(&'a self, new_child: &'a Node<'a, T>)

Append a new child to this node, after existing children.

pub fn prepend(&'a self, new_child: &'a Node<'a, T>)

Prepend a new child to this node, before existing children.

pub fn insert_after(&'a self, new_sibling: &'a Node<'a, T>)

Insert a new sibling after this node.

pub fn insert_before(&'a self, new_sibling: &'a Node<'a, T>)

Insert a new sibling before this node.