Module comrak::nodes

The CommonMark AST.

Structs

Ast
A single node in the CommonMark AST.

NodeCodeBlock
The metadata and data of a code block (fenced or indented).

NodeHeading
The metadata of a heading.

NodeHtmlBlock
The metadata of an included HTML block.

NodeLink
The details of a link's destination, or an image's source.

NodeList
The metadata of a list; the kind of list, the delimiter used and so on.

Enums

ListDelimType
The delimiter for ordered lists, i.e. the character which appears after each number.

ListType
The type of list.

NodeValue
The core AST node enum.

TableAlignment
Alignment of a single table cell.

Type Definitions

AstNode
The type of a node within the document.