Enum comrak::nodes::NodeValue

pub enum NodeValue {
    Document,
    BlockQuote,
    List(NodeList),
    Item(NodeList),
    CodeBlock(NodeCodeBlock),
    HtmlBlock(NodeHtmlBlock),
    Paragraph,
    Heading(NodeHeading),
    ThematicBreak,
    FootnoteDefinition(Vec<u8>),
    Table(Vec<TableAlignment>),
    TableRow(bool),
    TableCell,
    Text(Vec<u8>),
    SoftBreak,
    LineBreak,
    Code(Vec<u8>),
    HtmlInline(Vec<u8>),
    Emph,
    Strong,
    Strikethrough,
    Superscript,
    Link(NodeLink),
    Image(NodeLink),
    FootnoteReference(Vec<u8>),
}

The core AST node enum.

Variants
Document

The root of every CommonMark document. Contains blocks.

BlockQuote

Block. A block quote. Contains other blocks.

> A block quote.

List(NodeList)

Block. A list. Contains list items.

* An unordered list
* Another item

1. An ordered list
2. Another item

Item(NodeList)

Block. A list item. Contains other blocks.

CodeBlock(NodeCodeBlock)

Block. A code block; may be fenced or indented. Contains raw text which is not parsed as Markdown, although is HTML escaped.

HtmlBlock(NodeHtmlBlock)

Block. A HTML block. Contains raw text which is neither parsed as Markdown nor HTML escaped.

Paragraph

Block. A paragraph. Contains inlines.

Heading(NodeHeading)

Block. A heading; may be an ATX heading or a setext heading. Contains inlines.

ThematicBreak

Block. A thematic break. Has no children.

FootnoteDefinition(Vec<u8>)

Block. A footnote definition. The Vec<u8> is the footnote's name. Contains other blocks.

Table(Vec<TableAlignment>)

Block. A table per the GFM spec. Contains table rows.

TableRow(bool)

Block. A table row. The bool represents whether the row is the header row or not. Contains table cells.

TableCell

Block. A table cell. Contains inlines.

Text(Vec<u8>)

Inline. Textual content. All text in a document will be contained in a Text node.

SoftBreak

Inline. A soft line break. If the hardbreaks option is set in ComrakOptions during formatting, it will be formatted as a LineBreak.

LineBreak

Inline. A hard line break.

Code(Vec<u8>)

Inline. A code span.

HtmlInline(Vec<u8>)

Inline. Raw HTML contained inline.

Emph

Inline. Emphasised text.

Strong

Inline. Strong text.

Strikethrough

Inline. Strikethrough text per the GFM spec.

Superscript

Inline. Superscript. Enabled with ext_superscript option.

Link(NodeLink)

Inline. A link to some URL, with possible title.

Image(NodeLink)

Inline. An image.

FootnoteReference(Vec<u8>)

Inline. A footnote reference; the Vec<u8> is the referent footnote's name.

Methods

impl NodeValue

pub fn block(&self) -> bool

Indicates whether this node is a block node or inline node.

pub fn contains_inlines(&self) -> bool

Indicates whether this node may contain inlines.

pub fn text(&self) -> Option<&Vec<u8>>

Return a reference to the text of a Text inline, if this node is one.

Convenience method.

pub fn text_mut(&mut self) -> Option<&mut Vec<u8>>

Return a mutable reference to the text of a Text inline, if this node is one.

Convenience method.