Struct comrak::Arena

pub struct Arena<T> { /* fields omitted */ }

Methods

impl<T> Arena<T>

pub fn new() -> Arena<T>

pub fn with_capacity(n: usize) -> Arena<T>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn alloc(&self, value: T) -> &mut T

Allocates a value in the arena, and returns a mutable reference to that value.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn alloc_extend<I>(&self, iterable: I) -> &mut [T] where
I: IntoIterator<Item = T>,

Uses the contents of an iterator to allocate values in the arena. Returns a mutable slice that contains these values.

pub unsafe fn alloc_uninitialized(&self, num: usize) -> *mut [T]

Allocates space for a given number of values, but doesn't initialize it.

pub fn uninitialized_array(&self) -> *mut [T]

Returns unused space.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn into_vec(self) -> Vec<T>