Crate comrak

A 100% CommonMark and GFM compatible Markdown parser. Source repository is at https://github.com/kivikakk/comrak.

The design is based on cmark-gfm, so familiarity with that will help.

You can use comrak::markdown_to_html directly:

use comrak::{markdown_to_html, Options};
assert_eq!(markdown_to_html("Hello, **世界**!", &Options::default()),
           "<p>Hello, <strong>世界</strong>!</p>\n");

Or you can parse the input into an AST yourself, manipulate it, and then use your desired formatter:

use comrak::{Arena, parse_document, format_html, Options};
use comrak::nodes::{AstNode, NodeValue};

// The returned nodes are created in the supplied Arena, and are bound by its lifetime.
let arena = Arena::new();

let root = parse_document(
    &arena,
    "This is my input.\n\n1. Also my input.\n2. Certainly my input.\n",
    &Options::default());

fn iter_nodes<'a, F>(node: &'a AstNode<'a>, f: &F)
    where F : Fn(&'a AstNode<'a>) {
    f(node);
    for c in node.children() {
        iter_nodes(c, f);
    }
}

iter_nodes(root, &|node| {
    match &mut node.data.borrow_mut().value {
        &mut NodeValue::Text(ref mut text) => {
            let orig = std::mem::replace(text, String::new());
            *text = orig.replace("my", "your");
        }
        _ => (),
    }
});

let mut html = vec![];
format_html(root, &Options::default(), &mut html).unwrap();

assert_eq!(
    String::from_utf8(html).unwrap(),
    "<p>This is your input.</p>\n\
     <ol>\n\
     <li>Also your input.</li>\n\
     <li>Certainly your input.</li>\n\
     </ol>\n");

Re-exports

pub use html::format_document as format_html;

pub use html::format_document_with_plugins as format_html_with_plugins;

pub use html::Anchorizer;

Modules

adapters

Adapter traits for plugins.

arena_tree

Included from https://github.com/SimonSapin/rust-forest/blob/5783c8be8680b84c0438638bdee07d4e4aca40ac/arena-tree/lib.rs. MIT license (per Cargo.toml).

html

The HTML renderer for the CommonMark AST, as well as helper functions.

nodes

The CommonMark AST.

plugins

Plugins for enhancing the default implementation of comrak can be defined in this module.

Structs

Arena

An arena of objects of type T.

ExtensionOptions

Options to select extensions.

ExtensionOptionsBuilder

Builder for ExtensionOptions.

Options

Umbrella options struct.

ParseOptions

Options for parser functions.

ParseOptionsBuilder

Builder for ParseOptions.

Plugins

Umbrella plugins struct.

PluginsBuilder

Builder for Plugins.

RenderOptions

Options for formatter functions.

RenderOptionsBuilder

Builder for RenderOptions.

RenderPlugins

Plugins for alternative rendering.

RenderPluginsBuilder

Builder for RenderPlugins.

Enums

ListStyleType

Options for bulleted list redering in markdown. See link_style in RenderOptions for more details.

Functions

format_commonmark

Formats an AST as CommonMark, modified by the given options.

format_commonmark_with_plugins

Formats an AST as CommonMark, modified by the given options. Accepts custom plugins.

format_xml

Formats an AST as HTML, modified by the given options.

format_xml_with_plugins

Formats an AST as HTML, modified by the given options. Accepts custom plugins.

markdown_to_commonmark

Render Markdown back to CommonMark.

markdown_to_commonmark_xml

Render Markdown to CommonMark XML. See https://github.com/commonmark/commonmark-spec/blob/master/CommonMark.dtd.

markdown_to_commonmark_xml_with_plugins

Render Markdown to CommonMark XML using plugins. See https://github.com/commonmark/commonmark-spec/blob/master/CommonMark.dtd.

markdown_to_html

Render Markdown to HTML.

markdown_to_html_with_plugins

Render Markdown to HTML using plugins.

parse_document

Parse a Markdown document to an AST.

parse_document_with_broken_link_callback

Parse a Markdown document to an AST.

version

Return the version of the crate.

Type Definitions

ComrakExtensionOptions

Legacy naming of ExtensionOptions

ComrakOptions

Legacy naming of Options

ComrakParseOptions

Legacy naming of ParseOptions

ComrakPlugins

Legacy naming of Plugins

ComrakRenderOptions

Legacy naming of RenderOptions

ComrakRenderPlugins

Legacy naming of RenderPlugins