Function comrak::parse_document_with_broken_link_callback

pub fn parse_document_with_broken_link_callback<'a, 'c>(
    arena: &'a Arena<AstNode<'a>>,
    buffer: &str,
    options: &Options,
    callback: Option<&'c mut dyn FnMut(&str) -> Option<(String, String)>>
) -> &'a AstNode<'a>

Parse a Markdown document to an AST.

In case the parser encounters any potential links that have a broken reference (e.g [foo] when there is no [foo]: url entry at the bottom) the provided callback will be called with the reference name, and the returned pair will be used as the link destination and title if not None.

Note: The label provided to the callback is the normalized representation of the label as described in the GFM spec.

use comrak::{Arena, parse_document_with_broken_link_callback, format_html, Options};
use comrak::nodes::{AstNode, NodeValue};

// The returned nodes are created in the supplied Arena, and are bound by its lifetime.
let arena = Arena::new();

let root = parse_document_with_broken_link_callback(
    &arena,
    "# Cool input!\nWow look at this cool [link][foo]. A [broken link] renders as text.",
    &Options::default(),
    Some(&mut |link_ref: &str| match link_ref {
        "foo" => Some((
            "https://www.rust-lang.org/".to_string(),
            "The Rust Language".to_string(),
        )),
        _ => None,
    }),
);

let mut output = Vec::new();
format_html(root, &Options::default(), &mut output)?;
let output_str = std::str::from_utf8(&output).expect("invalid UTF-8");
assert_eq!(output_str, "<h1>Cool input!</h1>\n<p>Wow look at this cool \
                <a href=\"https://www.rust-lang.org/\" title=\"The Rust Language\">link</a>. \
                A [broken link] renders as text.</p>\n");